pyproject.toml
[project]
name = "snapcook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "imagehash>=4.3.2",
    "jinja2>=3.1.6",
    "load-dotenv>=0.1.0",
    "pillow>=12.0.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "uvicorn[standard]>=0.40.0",
]
